socksrv: do not wait for final EOF from R
Normally, ProxyHandler keeps a session alive until all data from the
remote end is consumed by a local listener, and if there is no connected
listener then it waits for one to connect. Previously this happened even
if the remote end did not send any more data, and the listener only had
to consume the final EOF. This is no longer the case.

 - Also remove the LIFNAME lookup until there's a cross-platform code.
 - Change log message whitespace formatting.
 - Separate SIGINT and SIGTERM handler functions.